requirements.txt
Django==2.1.3
djangorestframework==3.9.0
PyJWT==1.6.4
pytz==2018.7
